Make functions partial rather than asserting totality
@mjgpy3 I know why these are partial but I have not yet figured out how
to fix it. At this point, if we are to really take hold of the dependent
types we will have to depart from the examples on F# for Fun and Profit,
I believe. In any case, as currently implemented, they are definitely
partial and should not be asserted to be total.

I get an error on ['0'..'9'] saying that there is no Enum Char instance,
so I changed that.