Provide automated sequencing repair
There is no way I can renumber the sequencing manually after or during
the manual merge. I'm just going to do the manual merge after running
the sorting, then run this to repair the sequencing that I botched up in
the merged, then run the re-sorting...

Had to introduce a big fat component test to gain confidence in my
algorithm here. Also refactored on the introduction of the new code to
reduce duplication. Also ran into the limitations of having the sample
XML in its own free file and so got rid of that.